% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_BSA.R
\name{calculate_BSA}
\alias{calculate_BSA}
\title{Calculate body surface area (BSA) based on the Du Bois & Du Bois equation.}
\usage{
calculate_BSA(WT, HT)
}
\arguments{
\item{WT}{Body weight in kg.}

\item{HT}{Height in cm.}
}
\value{
Calculated BSA in m².
}
\description{
Calculate body surface area (BSA) based on the Du Bois & Du Bois equation.
}
\examples{
data$BSA <- calculate_BSA(data$WT, data$HT)
}
