% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_tree_model_ensembling_test.R
\name{classification_tree_model_ensembling_test}
\alias{classification_tree_model_ensembling_test}
\title{Test classification tree model ensembling to attribute patient individualized model weights.}
\usage{
classification_tree_model_ensembling_test(test_data, train_results)
}
\arguments{
\item{test_data}{Test dataset with ID to indicate a subject (set of covariates), covariates, observed target variable (for example CMAX), predicted target variable (for example CMAX_PRED), TIME if the target_variable is CONCENTRATION.}

\item{train_results}{Results of the algorithm training.}
}
\value{
The test data with the attributed model weights.
}
\description{
Test classification tree model ensembling to attribute patient individualized model weights.
}
\examples{
results <- classification_tree_model_ensembling_test(test_data = AMOX_CMAX_TEST, train_results = train_results)
}
