% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_tree_model_ensembling_train.R
\name{classification_tree_model_ensembling_train}
\alias{classification_tree_model_ensembling_train}
\title{Train classification tree informed model ensembling to calculate weights.}
\usage{
classification_tree_model_ensembling_train(
  data,
  target_variable,
  continuous_cov,
  categorical_cov,
  seed = 1991,
  conservative_pruning = FALSE
)
}
\arguments{
\item{data}{A dataset with ID to indicate a subject (set of covariates), covariates, observed target variable (for example CMAX), predicted target variable (for example CMAX_PRED), MODEL name used for the simulation.}

\item{target_variable}{A secondary PK parameter (CMAX - maximal concentration, AUC - exposure or CMIN - trough concentration).}

\item{continuous_cov}{List of continuous covariates.}

\item{categorical_cov}{List of categorical covariates.}

\item{seed}{Random seed for reproducibility.}

\item{conservative_pruning}{If FALSE (the default) take the highest CP with the lowest cross-validation error else take the highest CP whose cross-validation error is within 1 SD of the lowest cross-validation error.}
}
\value{
A table with the decision trees' leaves for different subgroups and models with the probability of correct predictions (prob_yes = model weight before standardization) and a list of rpart objects with the classification trees, plots of confusion matrices for each model.
}
\description{
Train classification tree informed model ensembling to calculate weights.
}
\examples{
leaf_results <- classification_tree_model_ensembling_train(data = data, target_variable = "CMAX", continuous_cov = c("WT", "CRCL"), categorical_cov = c("SEX", "BURN"))
}
