% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ft_over_mic.R
\name{compute_ft_over_mic}
\alias{compute_ft_over_mic}
\title{Compute fToverMIC between specified timepoints from mrgsolve simulation.}
\usage{
compute_ft_over_mic(sim_tbl, from = 0, to = 24)
}
\arguments{
\item{sim_tbl}{Tibble output of an mrgsolve simulation.}

\item{from}{Start of time interval over which fToverMIC will be computed.}

\item{to}{End of time interval over which fToverMIC will be computed.}
}
\value{
a tibble with added columns TOVERMIC and fT_over_mic. See details.
}
\description{
Compute fToverMIC between specified timepoints from mrgsolve simulation.
}
\details{
Make sure from and to are the same unit as time in the sim_tbl.
Column descriptions:
\itemize{
\item TOVERMIC absolute duration for which the unbound concentration is above
the MIC. Has the same units as the time column of sim_tbl
\item ft_over_mic fraction of the interval for which the unbound concentration is
above the minimal inhibitory concentration (MIC).
}
}
\examples{
compute_ft_over_mic(sim_tbl = sim_tbl, from = 24, to = 48)
}
