% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_performance_metrics.R
\name{compute_performance_metrics}
\alias{compute_performance_metrics}
\title{Calculate relative root mean squared error and relative mean prediction error for the test data results.}
\usage{
compute_performance_metrics(target_variable, test_data)
}
\arguments{
\item{target_variable}{A secondary PK parameter (CMAX - maximal concentration, AUC - exposure or CMIN - trough concentration, ft_above_MIC - time above minimal inhibitory concentration) or CONCENTRATION.}

\item{test_data}{test data containing the columns ID, WEIGHTED_PREDICTION, and depending on the target variable CMAX/CMIN/AUC or TIME and DV for CONCENTRATION.}
}
\value{
A tibble with rRMSE and MPE.
}
\description{
Calculate relative root mean squared error and relative mean prediction error for the test data results.
}
\examples{
compute_performance_metrics(test_data, target_variable = "CMAX")
}
