% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pta_t_above_mic.R
\name{compute_pta_t_above_mic}
\alias{compute_pta_t_above_mic}
\title{Compute probability of target attainment based on fToverMIC.}
\usage{
compute_pta_t_above_mic(tbl_t_over_mic, target)
}
\arguments{
\item{tbl_t_over_mic}{Tibble with columns ft_over_mic and MIC.}

\item{target}{Target fraction of time above MIC.}
}
\value{
Tibble with fraction of target attainment for each MIC value in
tbl_t_over_mic.
}
\description{
Compute probability of target attainment based on fToverMIC.
}
\examples{
compute_pta_t_above_mic(tbl_t_over_mic = tbl_t_over_mic, target = 1)
}
