% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_to_var.R
\name{cv_to_var}
\alias{cv_to_var}
\title{Calculate variance from variation coefficient of a parameter.}
\usage{
cv_to_var(distribution, mean = NULL, cv)
}
\arguments{
\item{distribution}{Parameter distribution across the population, can take values "normal" or "lognormal".}

\item{mean}{Parameter mean if the parameter is normally distributed.}

\item{cv}{Parameter's variation coefficient.}
}
\value{
The variance of the parameter.
}
\description{
Calculate variance from variation coefficient of a parameter.
}
\details{
Make sure to write "normal" or "lognormal" between quotation marks.
}
