% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_GFR.R
\name{estimate_GFR}
\alias{estimate_GFR}
\title{Estimate glomerular filtration rate (GFR) using the 2021 CKD-EPI equation.}
\usage{
estimate_GFR(AGE, CREAT, SEX)
}
\arguments{
\item{AGE}{Age in years.}

\item{CREAT}{Serum creatinine concentration in mg/dL.}

\item{SEX}{"M" or "F".}
}
\value{
Estimated GFR values in mL/min/1.73 m².
}
\description{
Estimate glomerular filtration rate (GFR) using the 2021 CKD-EPI equation.
}
\examples{
data$eGFR <- estimate_GFR(data$AGE, data$CREAT, data$SEX)
}
