% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famd_test.R
\name{famd_test}
\alias{famd_test}
\title{Factor Analysis of Mixed Data (FAMD) test to attribute model weights based on the similiarty of a patient's covariates to the model development cohorts}
\usage{
famd_test(test, train_results)
}
\arguments{
\item{test}{A test dataset with ID to indicate a subject (set of covariates), covariates, observed target variable (for example CMAX), predicted target variable (for example CMAX_PRED), MODEL name used for the simulation.}

\item{train_results}{The FAMD model resulting from the famd_train function.}
}
\value{
A goodness of fit (GOF) plot and the test dataset with the attributed weights and ensembled concentrations.
}
\description{
Factor Analysis of Mixed Data (FAMD) test to attribute model weights based on the similiarty of a patient's covariates to the model development cohorts
}
\examples{
result<- famd_train(test = test, train_results = train_results)
}
