% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famd_train.R
\name{famd_train}
\alias{famd_train}
\title{Factor Analysis of Mixed Data (FAMD) 'training' to attribute model weights based on the similiarty of a patient's covariates to the model development cohorts}
\usage{
famd_train(train, continuous_cov, categorical_cov, target_variable)
}
\arguments{
\item{train}{A training dataset with ID to indicate a subject (set of covariates), covariates, observed target variable (for example CMAX), predicted target variable (for example CMAX_PRED), MODEL name used for the simulation.}

\item{continuous_cov}{List of continuous covariates.}

\item{categorical_cov}{List of categorical covariates.}

\item{target_variable}{A secondary PK parameter (CMAX - maximal concentration, AUC - exposure or CMIN - trough concentration).}
}
\value{
The FAMD model with the principal components, centroids and covariance matrices.
}
\description{
Factor Analysis of Mixed Data (FAMD) 'training' to attribute model weights based on the similiarty of a patient's covariates to the model development cohorts
}
\examples{
result<- famd_train(train = train, continuous_cov = c("WT", "CREAT", "AGE"), categorical_cov = c("BURN", "SEX"), target_variable = "CMIN")
}
