% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_weight_plot.R
\name{generate_weight_plot}
\alias{generate_weight_plot}
\title{Plot to show average model weights for categories/quantiles of a covariate in the test data}
\usage{
generate_weight_plot(data, cov, n_quantiles = 4, is_categorical = FALSE)
}
\arguments{
\item{data}{The generated test dataset with the columns MODEL, WEIGHT and covariates}

\item{cov}{Covariate}

\item{n_quantiles}{If continuous covariate, to how many quantiles should it be divided to display weights. Default is 4.}

\item{is_categorical}{Is the covariate categorical? TRUE or FALSE, default is FALSE}
}
\value{
A plot showing avareage weights for each model in the covariates categories/quantiles
}
\description{
Plot to show average model weights for categories/quantiles of a covariate in the test data
}
\examples{
weight_BURN <- generate_weight_plot(test_data, "BURN", is_categorical = TRUE)
}
