% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_test.R
\name{knn_test}
\alias{knn_test}
\title{k nearest neighbors (KNN) for MIPD}
\usage{
knn_test(final_fit, test, seed = 1991)
}
\arguments{
\item{final_fit}{The trained KNN model from the knn_train function.}

\item{test}{Seed.}
}
\value{
A test data with containing the KNN predictions.
}
\description{
k nearest neighbors (KNN) for MIPD
}
\examples{
results <- knn_test(final_fit = final_fit, test = AMOX_CMIN_TEST)
}
