% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_train.R
\name{knn_train}
\alias{knn_train}
\title{k nearest neighbors (KNN) for MIPD}
\usage{
knn_train(train, continuous_cov, categorical_cov, grid_size = 10, seed = 1991)
}
\arguments{
\item{train}{A preprocessed (ml_data_preprocess function) training dataset with one line per patient containing covariates, the administered dose, and interdose interval.}

\item{continuous_cov}{List of continuous covariates.}

\item{categorical_cov}{List of categorical covariates.}

\item{grid_size}{Grid size for hyperparameter tuning.}

\item{seed}{Random seed.}
}
\value{
Trained KNN model, tuning autoplots, and tuned hyperparameters.
}
\description{
k nearest neighbors (KNN) for MIPD
}
\examples{
results <- knn_train(train = AMOX_CMIN_TRAIN, continuous_cov = c("WT", "CRCL"), categorical_cov = c("BURN", "OBESE"))
}
