% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{machine_learning}
\alias{machine_learning}
\title{Four machine learning models are fit to make predictions for the test data - Support Vector Machine (SVM), k nearest neighbors (KNN), XGBoost, Random forest (RF)}
\usage{
machine_learning(
  train,
  test,
  continuous_cov,
  categorical_cov,
  target_variable,
  target_concentration
)
}
\arguments{
\item{train}{A test dataset with one line per patient. The dataset should include the target variable, covariables and the DAILY dose.}

\item{continuous_cov}{List of continuous covariates.}

\item{categorical_cov}{List of categorical covariates.}

\item{target_variable}{A secondary PK parameter (CMAX - maximal concentration, AUC - exposure or CMIN - trough concentration).}

\item{target_concentration}{Target to reach (Cmax, Cmin or AUC).}
}
\value{
A test data with four additional columns containing the predictions for the four ML models.
}
\description{
Four machine learning models are fit to make predictions for the test data - Support Vector Machine (SVM), k nearest neighbors (KNN), XGBoost, Random forest (RF)
}
\examples{
results <- machine_learning(train = AMOX_CMIN_TRAIN, test = AMOX_CMIN_TEST, continuous_cov = c("WT", "CRCL"), categorical_cov = c("BURN", "OBESE"), target_variable = "CMIN", target_concentration = 60)
}
