% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_data_preprocess.R
\name{ml_data_preprocess}
\alias{ml_data_preprocess}
\title{Data preprocessing for machine learning}
\usage{
ml_data_preprocess(data, target_variable, target_concentration)
}
\arguments{
\item{data}{A dataset with one line per patient. The dataset should include the target variable (ending with _IND for example CMIN_IND), and the administered dose dose (DOSE_ADM) and a column called II for interdose interval (set to 24 for continuous administration).}

\item{target_variable}{A secondary PK parameter (CMAX - maximal concentration, AUC - exposure or CMIN - trough concentration).}

\item{target_concentration}{Target to reach (Cmax, Cmin or AUC).}
}
\value{
A test data the log transformed target dose, and dosing scheme based stratification.
}
\description{
Data preprocessing for machine learning
}
\examples{
results <- ml_data_preprocess(data = AMOX_CMIN_TRAIN, target_variable = "CMIN", target_concentration = 60)
}
