% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_tree_model_ensembling_test.R
\name{regression_tree_model_ensembling_test}
\alias{regression_tree_model_ensembling_test}
\title{Test regression tree model ensembling to attribute subjects individualized model weights}
\usage{
regression_tree_model_ensembling_test(test_data, train_results)
}
\arguments{
\item{test_data}{Test dataset with ID to indicate a subject (set of covariates), covariates, observed target variable (for example CMAX_IND), predicted target variable (for example CMAX_PRED), TIME if the target_variable is CONCENTRATION.}

\item{train_results}{Results of the classification tree training.}
}
\value{
The test data with the attributed model weights.
}
\description{
Test regression tree model ensembling to attribute subjects individualized model weights
}
\examples{
results <- regression_tree_model_ensembling_test(test_data = AMOX_CMAX_TEST, train_results = train_results)
}
