% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_tree_model_ensembling_train.R
\name{regression_tree_model_ensembling_train}
\alias{regression_tree_model_ensembling_train}
\title{Train regression tree informed model ensembling to calculate weights with log-transformed prediction/observation ratio as target variable}
\usage{
regression_tree_model_ensembling_train(
  data,
  target_variable,
  continuous_cov,
  categorical_cov,
  seed = 1991,
  conservative_pruning = FALSE
)
}
\arguments{
\item{data}{A dataset with ID to indicate a subject (set of covariates), covariates, observed target variable (for example CMAX), predicted target variable (for example CMAX_PRED), MODEL name used for the simulation.}

\item{target_variable}{A secondary PK parameter (CMAX - maximal concentration, AUC - exposure or CMIN - trough concentration, CONCENTRATION - complete concentration profileCONCEN).}

\item{continuous_cov}{List of continuous covariates.}

\item{categorical_cov}{List of categorical covariates.}

\item{seed}{Set seed for reproductibility.}

\item{conservative_pruning}{If FALSE (default) take the highest CP with the lowest cross-validation error else take the highest CP whose cross-validation error is within 1 SD of the lowest cross-validation error.}
}
\value{
Trained regression trees.
}
\description{
Train regression tree informed model ensembling to calculate weights with log-transformed prediction/observation ratio as target variable
}
\examples{
leaf_results <- regression_tree_model_ensembling_train(data = data, target_variable = "CMAX", continuous_cov = c("WT", "CRCL"), categorical_cov = c("SEX", "BURN"))
}
