% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_test.R
\name{rf_test}
\alias{rf_test}
\title{Random forest (RF) for MIPD}
\usage{
rf_test(final_fit, test, seed = 1991)
}
\arguments{
\item{final_fit}{The trained RF model from the rf_train function.}

\item{test}{A preprocessed (ml_data_preprocess function) training dataset with one line per patient.}

\item{seed}{Seed for reproducibility.}
}
\value{
A test data with containing the RF predictions.
}
\description{
Random forest (RF) for MIPD
}
\examples{
results <- rf_test(final_fit = final_fit, test = AMOX_CMIN_TEST)
}
