% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mrgsolve.R
\name{simulate_mrgsolve}
\alias{simulate_mrgsolve}
\title{Plot to show average model weights for categories/quantiles of a covariate in the test data}
\usage{
simulate_mrgsolve(
  selected_regimens,
  selected_IDs_dose,
  dosing_table,
  COV_table,
  observation_grid,
  model_list
)
}
\arguments{
\item{selected_IDs_dose}{List of which IDs (subjects) to include in the simulation.}

\item{dosing_table}{A dataset with the characteristics of the dosing regimens containing the following columns: regimen_id (dosing regimen identifier), time, tinf (duration of infusion), amt, ii (interdose interval), evid, cmt (compartment).}

\item{COV_table}{Is the covariate categorical? TRUE or FALSE, default is FALSE}

\item{observation_grid}{A list of time values for the observations.}

\item{model_list}{List of models to use for the simulation. The models' .cpp file should be provided by the user.}

\item{selected_regiments}{List of regimen_ids (dosing regimen identifiers)to use in the simulation.}
}
\value{
A dataset with PRED, IPRED & Y
}
\description{
Plot to show average model weights for categories/quantiles of a covariate in the test data
}
\examples{
simulated_set <- simulate_mrgsolve(selected_regimens = c(10000mg, 20000mg), selected_IDs_dose = c(1:2500), dosing_table = dosing_table, COV_table = Cov_table, observation_grid = seq(0, 24, by = 2), model_list = (model1, model2))
}
