% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_test.R
\name{svm_test}
\alias{svm_test}
\title{Support Vector Machine (SVM) for MIPD}
\usage{
svm_test(final_fit, test, seed = 1991)
}
\arguments{
\item{final_fit}{The trained SVM model from the svm_train function.}

\item{test}{A preprocessed (ml_data_preprocess function) training dataset with one line per subject.}

\item{seed}{Seed for reproducibility.}
}
\value{
A test data with containing the SVM predictions.
}
\description{
Support Vector Machine (SVM) for MIPD
}
\examples{
results <- svm_test(final_fit = final_fit, test = AMOX_CMIN_TEST)
}
