% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighed_model_ensembling_test.R
\name{weighed_model_ensembling_test}
\alias{weighed_model_ensembling_test}
\title{Test weighed model ensembling to attribute subjects individualized model weights}
\usage{
weighed_model_ensembling_test(test_data, train_results)
}
\arguments{
\item{test_data}{Test dataset with ID to indicate a subject (set of covariates), covariates, observed target variable (for example CMAX), predicted target variable (for example CMAX_PRED), TIME if the target_variable is CONCENTRATION.}

\item{train_results}{Results of the weighed model ensembling training.}
}
\value{
The test data with the attributed model weights and ensembled concentrations, goodnes of fit (GOF) plot and predicted/observed concentration ratio plot.
}
\description{
Test weighed model ensembling to attribute subjects individualized model weights
}
\examples{
result <- weighed_model_ensembling_test(train_results = train_results, test_data = AMOX_CMAX_TEST)
}
