% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighed_model_ensembling_train.R
\name{weighed_model_ensembling_train}
\alias{weighed_model_ensembling_train}
\title{Train weighed model ensembling to calculate model and subgroup influence scores}
\usage{
weighed_model_ensembling_train(
  data,
  target_variable,
  pen_RMSE = -25,
  pen_MPE = -3,
  subgroup_inf_lower_bound = 80,
  subgroup_inf_upper_bound = 125,
  continuous_cov,
  categorical_cov
)
}
\arguments{
\item{data}{A dataset with ID to indicate a subject (set of covariates), covariates, observed target variable (for example CMAX), predicted target variable (for example CMAX_PRED), and MODEL name used for the simulation.}

\item{target_variable}{A secondary PK parameter (CMAX - maximal concentration, AUC - exposure or CMIN - trough concentration), time above minimal inhibitory concentration (ft_above_MIC) or CONCENTRATION.}

\item{pen_RMSE}{\itemize{
\item Negative penalty for relative root mean squared error (RMSE), default = -25.
}}

\item{pen_MPE}{\itemize{
\item Negative penalty for relative mean prediction error (MPE), default = -3.
}}

\item{subgroup_inf_lower_bound}{Lower bound of the percentage interval within which a prediction is considered correct for subgroup influence calculation (Defaults to 80 \%).}

\item{subgroup_inf_upper_bound}{Upper bound of the percentage interval within which a prediction is considered correct for subgroup influence calculation (Defaults to 125 \%).}

\item{continuous_cov}{= List of continuous covariates.}

\item{categorical_cov}{= List of categorical covariates.}
}
\value{
A table with RMSE, MPE, model influence and subgroup influence for each model and covariate quantile/category.
}
\description{
Train weighed model ensembling to calculate model and subgroup influence scores
}
\examples{
weighed_model_ensembling_train(data = data, target_variable = "CMAX", pen_RMSE = -25, pen_MPE = -3, continuous_cov = c("WT", "CRCL"), categorical_cov = c("SEX", "BURN"))
}
