% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighed_model_ensembling_tuning.R
\name{weighed_model_ensembling_tuning}
\alias{weighed_model_ensembling_tuning}
\title{Automatic hyperparameter (RMSE and MPE penalty) tuning for weighed model ensembling. Selection criterion is target attainment in the training data.}
\usage{
weighed_model_ensembling_tuning(
  train_data,
  target_variable,
  continuous_cov,
  categorical_cov,
  penalties_grid = NULL,
  conc_inf = 40,
  conc_sup = 80,
  conc_target = 60,
  freq_column = "FREQ",
  cv_stratification_col = NULL,
  n_cv = 10,
  target_log_dose = TRUE
)
}
\arguments{
\item{train_data}{Training dataset with ID to indicate a subject (set of covariates), covariates, observed target variable (for example CMAX), predicted target variable (for example CMAX_PRED), TIME if the target_variable is CONCENTRATION.}

\item{target_variable}{A secondary PK parameter (CMAX - maximal concentration, AUC - exposure or CMIN - trough concentration) or CONCENTRATION.}

\item{continuous_cov}{List of continuous covariates.}

\item{categorical_cov}{List of categorical covariates.}

\item{penalties_grid}{Tibble with two columns pen_RMSE and pen_MPE containing penalties values to be tested.}

\item{conc_inf}{lower limit of target concentration interval.}

\item{conc_sup}{upper limit of target concentration interval.}

\item{conc_target}{target concentration.}

\item{freq_column}{name of the interdose interval column.}

\item{cv_stratification_col}{name of the column indicating the stratification levels for stratified cross-validation. Defaults to NULL.}

\item{n_cv}{number of cross validation folds. Defaults to 10.}

\item{target_log_dose}{should the tuning be based on log dose or dose ? Default to TRUE.}
}
\value{
Tuned rRMSE and MPE penalties.
}
\description{
Automatic hyperparameter (RMSE and MPE penalty) tuning for weighed model ensembling. Selection criterion is target attainment in the training data.
}
\examples{
best <- weighed_model_ensembling_tuning(train_data = AMOX_CMIN_TRAIN, test_data = AMOX_CMIN_TEST, target_variable = "CMIN", continuous_cov = c("WT", "CREAT", "AGE"),categorical_cov = c("OBESE", "ICU", "BURN", "SEX"), conc_inf = 40, conc_sup = 80, conc_target = 60, max_dose_g = 20, freq_column = "FREQ")
}
