% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_test.R
\name{xgb_test}
\alias{xgb_test}
\title{XGBoost (XGB) for MIPD}
\usage{
xgb_test(final_fit, test, seed = 1991)
}
\arguments{
\item{final_fit}{The trained XGBoost model from the xgb_train function.}

\item{test}{A preprocessed (ml_data_preprocess function) training dataset with one line per patient.}

\item{seed}{Seed for reproducibility.}
}
\value{
A test data with containing the XGBoost predictions.
}
\description{
XGBoost (XGB) for MIPD
}
\examples{
results <- xgb_test(final_fit = final_fit, test = AMOX_CMIN_TEST)
}
